platforms/gae_standard/py37/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
gevent==1.4.0
google-cloud-datastore
google-cloud-ndb
google-cloud-redis
redis
google-cloud-tasks
googleapis-common-protos
google-cloud-logging
uwsgi
meinheld
uvicorn==0.9
ujson
fastapi
falcon
aioify
